+++ vyre-libs/src/security/taint_kill.rs
@@ -17,7 +17,7 @@
 
 use vyre_foundation::ir::Program;
 use vyre_primitives::bitset::and_not::bitset_and_not;
-use vyre_primitives::graph::csr_forward_traverse::bitset_words;
+use vyre_primitives::bitset::bitset_words;
 
 pub(crate) const OP_ID: &str = "vyre-libs::security::taint_kill";